Add files via upload
The first type of clustering algorithm discussed in this course used the spatial distribution of points to determine cluster centers and membership. The most prominent implementation of this concept is the K-means cluster algorithm. This approach is conceptually simple and often fast, however, it requires knowledge of the number of clusters ahead of time. While there are automated methods for determining  𝑘  algorithmically, this requirement is still an impediment for some applications. An alternative, density-based clustering technique called Density-Based Spatial Clustering of Applications with Noise (DBSCAN) can be used instead.

The DBSCAN algorithm has several advantages over the K-means algorithm. First, DBSCAN automatically determines the number of clusters within a data set. Second, since the DBSCAN algorithm is a density-based clustering algorithm, the discovered clusters can have arbitrary shapes. On the other hand, since the clusters and their membership are defined by the density, the hyperparameters used to specify the target density can dramatically affect the cluster determination. Thus, hyperparameter tuning may be required to achieve optimal results.